requirements.txt
httpx
customtkinter
Pillow
packaging
pyinstaller
